abstract scroll logic + hooks into useGetScrollAttributes reusable hook
