+++ src/components/Hscroll/index.jsx
@@ -1,26 +1,7 @@
-import React, { useState, useRef } from 'react';
+import React from 'react';
 import cc from 'classcat';
 import PropTypes from 'prop-types';
-import rafSchd from 'raf-schd';
-
-/**
- * @function getScrollAtributes
- * Returns attributes about user scroll position.
- *
- * @param {Object} e - scroll event from overflow element
- * @param {Number} contentScrollWidth - scrollWidth of inner content element
- */
-export const getScrollAtributes = (e, contentScrollWidth) => {
-  const { scrollLeft, clientWidth } = e.target;
-
-  /* "end" of scroll is the total width minus the visible area */
-  const scrollEnd = contentScrollWidth - clientWidth;
-
-  return {
-    scrolled: Boolean(scrollLeft),
-    scrollEnd: Boolean(scrollLeft === scrollEnd),
-  };
-};
+import useGetScrollAttributes from './useGetScrollAttributes';
 
 /**
  * @param {Object} props react props
@@ -33,9 +14,7 @@ const Hscroll = ({
   children,
   ...rest
 }) => {
-  const [isScrolled, setIsScrolled] = useState(false);
-  const [isAtScrollEnd, setIsAtScrollEnd] = useState(false);
-  const contentEl = useRef(null);
+  const [onScroll, contentEl, isScrolled, isAtScrollEnd] = useGetScrollAttributes();
 
   const containerClassNames = cc([
     {
@@ -47,12 +26,6 @@ const Hscroll = ({
     'hscroll',
     `hscroll--bgColor--${bgColor}`,
   ]);
-
-  const onScroll = (e) => {
-    const { scrolled, scrollEnd } = getScrollAtributes(e, contentEl.current.scrollWidth);
-    rafSchd(setIsScrolled(scrolled));
-    rafSchd(setIsAtScrollEnd(scrollEnd));
-  };
 
   return (
     <div
